+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Orschlurch channel
+
 ### Changed
 - Event channel size extracted into properties variable (`grab.channel-size`)